+++ src/components/DocumentDetail/index.jsx
@@ -1,0 +1,7 @@
+import * as s from './DocumentDetail.style';
+
+const DocumentDetail = () => {
+  return;
+};
+
+export default DocumentDetail;
